// src/features/events/model/app-proto/sip.schema.ts
import { z } from 'zod';

import { baseEventSchema } from '@/features/events/model/event';

export const sipSchema = z.object({
  method: z.string(),
  request_line: z.string(),
  uri: z.string(),
  version: z.string(),
});

export const sipEventSchema = baseEventSchema.extend({
  event_type: z.literal('sip'),
  app_proto: z.literal('sip'),
  sip: sipSchema,
});

export type SipEvent = z.infer<typeof sipEventSchema>;